backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
pytest==8.0.0
pytest-django==4.7.0
